refactor(ui): modularize App.tsx into hooks and components (Jules session 2069836845276372697)
