+++ custom_components/meraki_ha/www/src/types/ha-frontend.d.ts
@@ -8,7 +8,7 @@ declare global {
       'ha-settings-row': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { narrow?: boolean; slot?: string }, HTMLElement>;
       'ha-switch': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { checked?: boolean; disabled?: boolean }, HTMLElement>;
       'ha-circular-progress': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { active?: boolean; alt?: string }, HTMLElement>;
-      'ha-tabs': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { selected?: string; "attr-for-selected"?: string; scrollable?: boolean }, HTMLElement>;
+      'ha-tabs': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { selected?: string; "attr-for-selected"?: string; scrollable?: boolean; "onIron-select"?: (e: CustomEvent) => void }, HTMLElement>;
       'paper-tab': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { name?: string }, HTMLElement>;
       'ha-textfield': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { label?: string; value?: string | number; placeholder?: string; disabled?: boolean; type?: string; icon?: string; iconTrailing?: string; required?: boolean; errorMessage?: string; validationMessage?: string; }, HTMLElement>;
       'ha-select': React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement> & { label?: string; value?: string; disabled?: boolean; fixedMenuPosition?: boolean; naturalMenuWidth?: boolean; required?: boolean; }, HTMLElement>;
@@ -18,4 +18,7 @@ declare global {
     }
   }
 }
-export {};
+
+export interface HaTabsElement extends HTMLElement {
+  selected: string;
+}
